requirements.txt
libclang>=16.0.0
clang>=16.0.0
requests>=2.31.0
PyYAML>=6.0
